requirements.txt
absl-py==2.2.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
black==23.11.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
coverage==7.8.2
cycler==0.12.1
dash==3.0.4
dataclasses==0.6
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docker-pycreds==0.4.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-auth-oauthlib==1.2.2
grpcio==1.71.0
h11==0.16.0
h5py==3.10.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.40.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbdime==3.2.1
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
notebook==7.4.2
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.2.2
open3d==0.17.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==5.17.0
pluggy==1.6.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.23.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydicom==3.0.1
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyvoxel==0.0.2
PyYAML==6.0
pyzmq==26.4.0
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.15.3
Send2Trash==1.8.3
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.15.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
torch==2.7.0
torch_cluster==1.6.3
torch_geometric==2.5.0
torch_scatter==2.1.2
torch_sparse==0.6.18
torch_spline_conv==1.2.2
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.66.0
traitlets==5.14.3
trimesh==4.0.4
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
wandb==0.15.12
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
yarl==1.20.0
zipp==3.21.0
